Added new CQRS-like structure to most features
