+++ Fitz/Features/Accounts/Commands/AccountSlashCommands.cs
@@ -51,76 +51,6 @@ namespace Fitz.Features.Accounts
             #endregion Account Creation Interactions
         }
 
-        #region Signup
-
-        [SlashCommand("signup", "Just sign this form.")]
-        public async Task Signup(InteractionContext ctx)
-        {
-            Result accountCreationResult = await accountService.CreateAccountAsync(ctx.User);
-            if (accountCreationResult.Success)
-            {
-                Account account = accountCreationResult.Data as Account;
-                await this.bankService.AwardAccountCreationBonusAsync(account);
-
-                await ctx.CreateResponseAsync(DiscordInteractionResponseType.ChannelMessageWithSource,
-                    new DiscordInteractionResponseBuilder()
-                    .AddEmbed(accountEmbed(ctx.User, account))
-                    .AsEphemeral(true));
-
-                // check to see if the user is in the Waterbear guild
-                if (ctx.Guild.Id == Guilds.Waterbear)
-                {
-                    // assign a new role to a user
-                    await ctx.Guild.GetMemberAsync(ctx.User.Id).Result.GrantRoleAsync(ctx.Guild.GetRole(Roles.Accounts));
-                    return;
-                }
-                else
-                {
-                    DiscordGuild guild = await ctx.Client.GetGuildAsync(Guilds.Waterbear);
-                    DiscordMember discordMember = await guild.GetMemberAsync(ctx.User.Id);
-
-                    // check to see if the user is in the guild
-                    if (discordMember == null)
-                    {
-                        await ctx.CreateResponseAsync(DiscordInteractionResponseType.ChannelMessageWithSource, new DiscordInteractionResponseBuilder().WithContent($"You need to be in the Waterbear guild to get the Accounts role."));
-                        return;
-                    }
-                    else
-                    {
-                        await discordMember.GrantRoleAsync(guild.GetRole(Roles.Accounts));
-                        return;
-                    }
-                }
-            }
-            else
-            {
-                await ctx.CreateResponseAsync(DiscordInteractionResponseType.ChannelMessageWithSource, new DiscordInteractionResponseBuilder().WithContent($"You already have an account.").AsEphemeral(true));
-            }
-        }
-
-        #endregion Signup
-
-        #region Profile
-
-        [SlashCommand("profile", "You're hired.")]
-        [RequireAccount]
-        public async Task Profile(InteractionContext ctx)
-        {
-            Account account = accountService.FindAccount(ctx.User.Id);
-            if (account != null)
-            {
-                await ctx.CreateResponseAsync(DiscordInteractionResponseType.ChannelMessageWithSource,
-                    new DiscordInteractionResponseBuilder().AddEmbed(accountEmbed(ctx.User, account)).AsEphemeral(true));
-            }
-            else
-            {
-                await ctx.CreateResponseAsync(DiscordInteractionResponseType.ChannelMessageWithSource,
-                    new DiscordInteractionResponseBuilder().WithContent($"Doesn't seem like you have an account. Try running `/signup`.").AsEphemeral(true));
-            }
-        }
-
-        #endregion Profile
-
         #region Profile Lookup
 
         [SlashCommand("lookup", "Checkout someone elses profile.")]
