+++ Fitz/Features/Profile/GetProfile/Discord/ProfileAdapter.cs
@@ -1,0 +1,45 @@
+using DSharpPlus;
+using DSharpPlus.Entities;
+using DSharpPlus.EventArgs;
+using DSharpPlus.ModalCommands;
+using DSharpPlus.SlashCommands;
+using Fitz.Core.Commands.Attributes;
+using Fitz.Core.Contexts;
+using Fitz.Core.Models;
+using Fitz.Features.Accounts.Models;
+using Fitz.Features.Bank;
+using Fitz.Features.Lottery;
+using Fitz.Features.Lottery.Models;
+using Fitz.Features.Polls;
+using Fitz.Features.Polls.Models;
+using Fitz.Features.Rename;
+using Fitz.Variables;
+using Fitz.Variables.Emojis;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Fitz.Features.Profiles
+{
+    [SlashModuleLifespan(SlashModuleLifespan.Scoped)]
+    public sealed class ProfileAdapter : ApplicationCommandModule
+    {
+        [SlashCommand("profile", "You're hired.")]
+        [RequireAccount]
+        public async Task Profile(InteractionContext ctx)
+        {
+            Account account = accountService.FindAccount(ctx.User.Id);
+            if (account != null)
+            {
+                await ctx.CreateResponseAsync(DiscordInteractionResponseType.ChannelMessageWithSource,
+                    new DiscordInteractionResponseBuilder().AddEmbed(accountEmbed(ctx.User, account)).AsEphemeral(true));
+            }
+            else
+            {
+                await ctx.CreateResponseAsync(DiscordInteractionResponseType.ChannelMessageWithSource,
+                    new DiscordInteractionResponseBuilder().WithContent($"Doesn't seem like you have an account. Try running `/signup`.").AsEphemeral(true));
+            }
+        }
+    }
+}
